Use address instead of publicKey for the Census
I made a mistake in the naming. We'll need to change the form/fields/etc
accordingly, as what we need to send to Vocdoni API is the Wallet's Address
and not the Public Key.

Also we refactor a bit this function and introduce async/await, as
the getAddress method from ethers.js Wallet returns a Promise.