Fix bug with multiple elections in the same component
Until now it takes only the first election in the component, ignoring
the rest. With this fix, it'll take into account only the electionId of
the election that we want to create.

We still pass the componentId as I couldn't find the way of skipping
that, but as the electionId is unique in a Decidim installation, it
makes sense that we could skip the componentId too (given the correct
graphQL query).