Change JavaScript logic of window.Decidim.setupVoteComponent
Until now the PreviewVoteComponent and VoteComponent were retrieved from
the window.Decidim.setupVoteComponent object, and this worked well with
the Decidim approach of having a full page for login (aka census
checking), but this doesn't play nice with the full JavaScript approach
of the Vocdoni SDK, so I needed to change the mount logic.

This also means that we change the logic on when to use the Preview (Fake) voting:
until now this was done by just adding whichever script you needed
(setup-vote.js or setup-preview.js); now we must do it with the
`data-preview` attribute, as that allows us to send the wallet object to
the VoteComponent object in setup-vote.js.

This has implications also with the Census check. If it's on Preview
mode we skip it altogether, but if not we use the isInCensus method from
the Vocdoni SDK to check this.