Add "Setup Election" step in admin  (#18)
* Lock to spring-watcher-listen gem patch release

* Bump decidim from 0.27.0 to 0.27.1

* Question & Answers administration

* Add GraphQL API types and specs

* Answers
* Questions
* Election
* Elections

* Update repository in README

* Add Steps controller in admin

* Add missing db directory in gem specification

* Add webpacker configuration for JavaScript

* Add 'Create Wallet' logic in admin

* Use Ethereum Faker in specs

* Don't allow access to Wallet creation if it's already created

* Add 'description' field in Questions

* Add Questions and Answers to seeds

* Fix typo

* Fix typo

* Add 'npm ci' step in GitHub Action workflow CI conf

* Add 'Create Election' step in Admin

* Don't allow editing an Election if it's blocked

* Add 'Create Wallet' system spec

* Normalize i18n files with i18n-tasks

* Add elections counter in admin's component sidebar

* Add missing query for admin's component sidebar

* Add system spec for admin's elections steps controller

* Add generated package-lock.json

* Add missing preffix for Voter factories

* Fix eslint warnings in setup_election.js

* Fix bug with multiple elections in the same component

Until now it takes only the first election in the component, ignoring
the rest. With this fix, it'll take into account only the electionId of
the election that we want to create.

We still pass the componentId as I couldn't find the way of skipping
that, but as the electionId is unique in a Decidim installation, it
makes sense that we could skip the componentId too (given the correct
graphQL query).

* Remove misleading comment about multiple question support

Sometime ago, I think we only sent one question to Vocdoni API.

I've justed checked it and we sent all the questions (with answers)
correctly.