Add Census controller in Admin (#20)
* Lock to spring-watcher-listen gem patch release

* Bump decidim from 0.27.0 to 0.27.1

* Question & Answers administration

* Add GraphQL API types and specs

* Answers
* Questions
* Election
* Elections

* Update repository in README

* Add Steps controller in admin

* Add missing db directory in gem specification

* Add webpacker configuration for JavaScript

* Add 'Create Wallet' logic in admin

* Use Ethereum Faker in specs

* Don't allow access to Wallet creation if it's already created

* Add 'description' field in Questions

* Add Questions and Answers to seeds

* Fix typo

* Fix typo

* Add 'npm ci' step in GitHub Action workflow CI conf

* Add 'Create Election' step in Admin

* Don't allow editing an Election if it's blocked

* Add 'Create Wallet' system spec

* Normalize i18n files with i18n-tasks

* Add elections counter in admin's component sidebar

* Add missing query for admin's component sidebar

* Add system spec for admin's elections steps controller

* Add generated package-lock.json

* Add Census controller in Admin

* Add system spec

* Add credentials/wallets generation for Census

* Rename CsvDatum model to Voter

* Rename 'csv_data' table to 'voters' and unify migrations

* Add missing frozen_string_literal in migration file

* Add VocdoniVotersType to GraphQL API schema

* Add census logic check in SetupForm

* Use address instead of publicKey for the Census

I made a mistake in the naming. We'll need to change the form/fields/etc
accordingly, as what we need to send to Vocdoni API is the Wallet's Address
and not the Public Key.

Also we refactor a bit this function and introduce async/await, as
the getAddress method from ethers.js Wallet returns a Promise.

* Add validation and types to CensusCredentialForm

* Use the Wallets addresses for Census

Remove the Census textaraea for the demo and use the addresses provided
by the GraphQL API.

* Call explicitly to run method in SetupElection

* Replace wallet "public key" references for "address"

I messed up the naming and what we need to send to Vocdoni API are
addresses, so it's better to rename it accordingly.

* Add missing preffix for Voter factory

* Add 'npm ci' step in GitHub Action workflow CI conf

* Fix bad merge resolutions

* Solve bad git merge

* Bump @vocdoni/sdk package from v0.0.2-alpha to v0.0.3-alpha

* Add deterministic wallet generation for Census

* Add missing preffix for Voter factories

* Add missing preffix for Voter factories

* Add missing attribute from Question factory

This probably comes from a bad merge

* Add missing preffix for Voter factories

* Add missing i18n-tasks gem and spec

* Run i18n-tasks add-missing

* Ignore false detections of i18n-tasks 'unused' command

* Run i18n-tasks normalize

* Fix eslint warnings in setup_election.js

* Add debug message to i18n-tasks unused spec

* Revert "Add debug message to i18n-tasks unused spec"

This reverts commit 82efaaa29ac2dd898992c64b5c9f6a9da0818f8b.

* Remove unused i18n string

* Fix bug with multiple elections in the same component

Until now it takes only the first election in the component, ignoring
the rest. With this fix, it'll take into account only the electionId of
the election that we want to create.

We still pass the componentId as I couldn't find the way of skipping
that, but as the electionId is unique in a Decidim installation, it
makes sense that we could skip the componentId too (given the correct
graphQL query).

* Remove misleading comment about multiple question support

Sometime ago, I think we only sent one question to Vocdoni API.

I've justed checked it and we sent all the questions (with answers)
correctly.

* Resolve bad git merge conflict resolutions

* Resolve bad git merge conflict resolutions (part 2)

* Add time_before check on Election Setup

This is extracted from Elections module, but in our case we change it to
minutes instead of hours, as that was a problem that we had while
working on that module, as we had to wait 1 hour between checks.

* Add missing PlainCensus object (from bad merge)

* Add missing console.log debug message for electionId

* Run i18n-tasks normalize

* Fix linter offenses and remove dead code from bad merge

* Fix rubocop offenses

* Replaces usage of OpenStruct with a Class

* Add a stub and example for :setup_minimum_minutes_before_start in spec

* Fix message when generating census credentials

* Run i18n-tasks normalize