Change election step to "vote_ended" even if the step wasn't "vote"
This is something that came from a Code Review, where a developer was
trying the voting and they forgot to run the task during the "vote" step,
so they needed to change the step manually to "vote_ended" once the
"end_time" has arrived.
If the election is already on "created" step, then it's OK and there
shouldn't be any problem changing the step directly to "vote_ended".