Change "available credits" logic
While developing the `interruptible` feature, I've seen that the current
approach didn't fully worked, as we actually need to use credits while
doing these actions (cancel, pause, end, and continue). This means that
I needed to show the "available credits" on multiple pages, and also use
the .collectFaucetTokens when doing those actions.

As this is the Vocdoni monetization strategy, I think its better to make
this action explicit, even on development/staging, so administrators are
aware of this future expense.