Fix and improve details on the setup and voting flow (part II) (#42)
* Add "Results published" step in admin

* Remove unecessary :tally_started and :tally_ended steps

* Remove unecessary :tally_ended step partial

* Run i18n-tasks remove-unused

* Fix eslint offense

* Fix rubocop offense

* Add Q&A in an Election public page

* Remove results from spec

* Add multiple questions to factory

* Add :simple trait to Election's factory

If not we would need to radically change the preview_vote_online_spec
system's spec with multiple questions.

* Improve seeds with different Election statuses

* Add Results in an Election public page

* Add "none" as a status on Election's seeds

"none" isn't actually an status, but we need something to represent
the election not being created yet in the Vocodni API.

This allows us to have something almost ready to finish and sent to the
API.

* Fix bug on Election trait in factory

* Remove unecessary attachments on GraphQL API example

* Refactor SetupForm secrets on Steps

* Change the Census generation flow

After feedback from a demo, we've noticed that the flow isn't clear
enough, so we change how we present the Census upload and Wallets
generation flow.

Mostly, we only show the minimum information for each step, and we
change the texts/buttons, from "Generate credentials" to "Confirm the
census data"

* Remove wallet check from example

* Fix erblint offense

* View the number of credits available in Admin's steps

* Change 'Stream uri' to 'Video stream link'

Also, move it to the bottom to better signal that's an optional field.

* Show breadcrumb in Admin's steps

* Drop the "(Optional)" from Gallery label form

* Change "Elections" to "Elections (Vocdoni)" in the Admin's title

* Remove check for component and participator_space published on SetupForm

* Add "Fix it" links to SetupForm errors

So it's easier to admins to find out where they can solve the error.

* Reorder actions in Admin's elections list

* Add credits label to secondary in 'Create election' step

* Add Vocdoni logo on Admin's steps

* Add Vocdoni logo on Voting booth

* Fix bug on seeds with blocked Elections

* Add secretUntilTheEnd setting on Election

This setting configures on the Vocdoni API if the election has to be
secret until the end date. If not, the results are visible in real-time.

* Refactor Vocdoni SDK client initialization on Steps

* Rename SetupElection JS class to CreateVocdoniElection

* Be consistent when using the vocdoniElectionId variable name

* Fix bug on printing on console.log the electionId

* Rename setVocdoniClient() method to _setVocdoniClient()

* Replace 'vote sent' with 'vote cast'

* Replace 'check status' with 'access'

* Improve description on admin's elections' index

* Improve 'powered by' footer on admin's steps' page

* Run i18n-tasks normalize

* Change vote verification step on Voting booth

Most notably:

* Change the copies
* Change "vote id" to "vote receipt"
* Add the "copy" button
* Change "check my vote" to "verify my vote"

* Increase vote preview time to make it more realistic

* Change vocdoni logo extension from svg to png

* Fix vocdoni logo UI in Admin's steps footer

* Add description in 'Setup election' step and actions on 'Election created' step

* Extract action buttons for 'Preview landing page' and 'View at Explorer vote' in admin steps

* Replace vocdoni logo for new version

* Fix specs

* Add the election initialization to the exception handling

* Add the election_types attributes to the seeds

* Fix eslint offenses

* Fix spec

* Fix bug with setup election when the default locale isn't English

* Remove unused options (leftover from initVocdoniClient extraction)

* Fix bug when changing elections' steps without status

* Fix bug when accessing directly to the voting booth by URL

* Validate if the results are not empty

This partially mitigates a bug when we didn't fetch the results.
If all the results are 0, that means that we should show an error.

* Show an error when we couldn't fetch the results

This is done because when we have the "secret_until_the_end" setting
enabled. Even though the end time of the Election has arrived, the
Vocdoni SDK doesn't publish the results on time, so you can have errors
fetching them.

* Change button copy from View to Preview

As we're reusing this button on the "created" and "vote" steps, it just
makes more sense to always talk about "viewing" the Election page.

* Change election step to "vote_ended" even if the step wasn't "vote"

This is something that came from a Code Review, where a developer was
trying the voting and they forgot to run the task during the "vote" step,
so they needed to change the step manually to "vote_ended" once the
"end_time" has arrived.
If the election is already on "created" step, then it's OK and there
shouldn't be any problem changing the step directly to "vote_ended".

* Fix eslint offenses

* Fix rubocop offense

* Fix erblint offense

* Fix rubocop offenses

* Rename user to admin in rspec examples

* Update config/locales/en.yml

Co-authored-by: Andrés Pereira de Lucena <andreslucena@users.noreply.github.com>

---------

Co-authored-by: Andrés Pereira de Lucena <andreslucena@gmail.com>
Co-authored-by: Andrés Pereira de Lucena <andreslucena@users.noreply.github.com>