Add `value` attribute to Answers
Until now we've used the `id` as a value for the Answers. The problem is
that starting with v0.0.4-alpha of the SDK, we need to send the Vote
with the same order of the Answers that were sent during the setup, so
to make it cleaner and not introduce iconsistencies down the line, we
save the Answers values during the Setup and use this same value to send
the Vote.