Add a workaround for the cases where the client.url is not set
We need to set it manually, depending on the environment

This is something that happened to other people on the development team,
but that I couldn't reproduce on my machine. Starting with v0.0.4-alpha,
I've catched this a couple of times, so I prefer to add this workaround
until we have more information on what/why is happening.

We should pinpoint what exactly is the culprit of this issue and if it's
a bug in the SDK, we should fix it there and update.