Bump @vocdoni/sdk package from v0.0.3-alpha to v0.0.4-alpha (#45)
* Bump @vocdoni/sdk package from v0.0.3-alpha to v0.0.4-alpha

* Drop validation for at least one photo/attachment on Election Setup

From the vocdoni-sdk v0.0.4-alpha CHANGELOG:

> header and streamUri are no longer mandatory when creating an election.

* Add `value` attribute to Answers

Until now we've used the `id` as a value for the Answers. The problem is
that starting with v0.0.4-alpha of the SDK, we need to send the Vote
with the same order of the Answers that were sent during the setup, so
to make it cleaner and not introduce iconsistencies down the line, we
save the Answers values during the Setup and use this same value to send
the Vote.

* Add a workaround for the cases where the client.url is not set

We need to set it manually, depending on the environment

This is something that happened to other people on the development team,
but that I couldn't reproduce on my machine. Starting with v0.0.4-alpha,
I've catched this a couple of times, so I prefer to add this workaround
until we have more information on what/why is happening.

We should pinpoint what exactly is the culprit of this issue and if it's
a bug in the SDK, we should fix it there and update.

* Show an error message when `maxVoteOverwrites` has reached

* Fix rubocop offenses

* Run i18n-tasks normalize