Implement 'interruptible' election setting (#47)
* Add HTML titles to admin pages

* Use Decidim::Vocdoni.config for accessing module configuration

* Add interruptible setting on Election

This setting configures on the Vocdoni API if the election can be
paused, canceled or ended before the end_time.
In case it was paused, it can be continued.
Canceled and ended are irreversible.

* Add visual feebaack (spinner) while doing AJAX operations

* Change "available credits" logic

While developing the `interruptible` feature, I've seen that the current
approach didn't fully worked, as we actually need to use credits while
doing these actions (cancel, pause, end, and continue). This means that
I needed to show the "available credits" on multiple pages, and also use
the .collectFaucetTokens when doing those actions.

As this is the Vocdoni monetization strategy, I think its better to make
this action explicit, even on development/staging, so administrators are
aware of this future expense.

* Add "canceled" and "paused" status on Election

* Add "interruptible" logic

If this checkbox is enabled, then we show a "Danger zone" in the
admin panel once the Election is in the "vote" step. This allows the
admin to:

* Cancel an election: doesn't publish the results
* End an election: publish the results. It's the same result as if the
  election has reached the configured end_time.
* Pause an election: momentarily doesn't accept votes

* Show a message when logging if the election is paused

* Run i18n-tasks normalize

* Handle canceled/paused elections voting frontend

* Add verbosity and quiet argument to :change_election_status task

* Fix CSS for questions previews' accordions

* Make elections frontend aware of "canceled" and "paused" status

* Fix bug when trying to get the Votes and there isn't any Election

* Fix bug when checking the ElectionStatus while login

* Fix bug when passing the Wallet for voting

* Run i18n-tasks normalize

* Fix typo

* Always create a new wallet on Steps spec

* Change default Vocodni API endpoint to staging

* Make the Census email case insensitive

* Prefer usage of `env` over `vocdoniEnv` in JS variable names

* Fix hardcoding Explorer Vote dev environment URL on 'vote verify'

* Fix accordion CSS on vote stats and results

* Fix bug while publishing results

* Show styles (e.g. line breaks) on Election description

* Fix permission when election has finished prematurely (canceled or manually ended)

* Fix bug on focus mode when changing the Organizations' primary color

* Fix external link indicator icon in Vocodni logo

* Fix bug on `Copy` button when changing the Organizations' primary color

There's a bug on Decidim related with the Copy button on Share modal.
I proposed a fix and when that's accepted, backported and released,
we can remove this rule.

The fix is proposed on https://github.com/decidim/decidim/pull/10526

* Change copies on 'Vote successfully casted' page

* Add more specs and improve the ongoing/finished methods on Elections

* Make the 'answers values' set-up transparent to Admins

As it's an implementation detail, it doesn't make much sense to have it
an an explicit option during the Election Setup Step.

We do it with AJAX just before the Setup.

* Fix eslint offenses

* Add error details when Election Setup failed

* Remove 'Available credits' expectation in system spec

* Improve error logging during Setup Election step

* Fix Setup Election spec by waiting for the account to be created in the API

* Undo rubocop change on Election status changer task output

* fix codecov

* Drop quiet argument from change_election_status task

Suggested by code review. It's not necessary as we can redirect output
to /dev/null on the crontab. (> /dev/null)

* Get the answers_values_election_path from a data attribute.

Suggested by code review. It's much more cleaner and less buggy.

* Fix erblint offense

---------

Co-authored-by: Ivan Vergés <ivan@pokecode.net>